Moved content blocks to designated folder
