+++ src/ui/content_blocks/block/block_element.css
