+++ src/ui/content_blocks/code_editors/code_editors_element.js
@@ -1,4 +1,4 @@
-import BlockElement from 'src/ui/block/block_element'
+import BlockElement from 'src/ui/content_blocks/block/block_element'
 
 class CodeEditorsElement extends BlockElement {
 	get mediaSource() { return '/db/code_editors/code_editors.json' }
