+++ src/ui/content_blocks/conferences/conferences_element.js
@@ -1,4 +1,4 @@
-import BlockElement from 'src/ui/block/block_element'
+import BlockElement from 'src/ui/content_blocks/block/block_element'
 
 class ConferencesElement extends BlockElement {
 	get mediaSource() { return '/db/conferences/conferences.json' }
