+++ src/ui/content_blocks/dev_browsers/dev_browsers_element.js
@@ -1,9 +1,9 @@
-import BlockElement from 'src/ui/block/block_element'
+import BlockElement from 'src/ui/content_blocks/block/block_element'
 
-class BuildSystemsElement extends BlockElement {
-	get mediaSource() { return '/db/build_systems/build_systems.json' }
+class DevBrowsersElement extends BlockElement {
+	get mediaSource() { return '/db/dev_browsers/dev_browsers.json' }
 	get mediaPictureExtension() { return 'png' }
-	get mediaType() { return 'Build system' }
+	get mediaType() { return 'Browser' }
 
 	generateItem(item) {
 		let image = this.data.media.images+item.id+'.'+this.mediaPictureExtension
@@ -20,4 +20,4 @@ class BuildSystemsElement extends BlockElement {
 	}
 }
 
-export default BuildSystemsElement
+export default DevBrowsersElement
