+++ src/ui/content_blocks/libraries/libraries_element.css
