+++ src/ui/content_blocks/online_training/online_training_element.css
