+++ src/ui/registry.js
@@ -1,14 +1,14 @@
 import HeaderElement from 'src/ui/header/header_element'
-import BooksElement from 'src/ui/books/books_element'
-import CodeEditorsElement from 'src/ui/code_editors/code_editors_element'
-import FrameworksElement from 'src/ui/frameworks/frameworks_element'
-import ToolsElement from 'src/ui/tools/tools_element'
-import OnlineTrainingElement from 'src/ui/online_training/online_training_element'
-import DevBrowsersElement from 'src/ui/dev_browsers/dev_browsers_element'
-import BuildSystemsElement from "src/ui/build_systems/build_systems_element"
-import ConferencesElement from "src/ui/conferences/conferences_element"
-import ArticlesElement from "src/ui/articles/articles_element"
-import LibrariesElement from "src/ui/libraries/libraries_element"
+import BooksElement from 'src/ui/content_blocks/books/books_element'
+import CodeEditorsElement from 'src/ui/content_blocks/code_editors/code_editors_element'
+import FrameworksElement from 'src/ui/content_blocks/frameworks/frameworks_element'
+import ToolsElement from 'src/ui/content_blocks/tools/tools_element'
+import OnlineTrainingElement from 'src/ui/content_blocks/online_training/online_training_element'
+import DevBrowsersElement from 'src/ui/content_blocks/dev_browsers/dev_browsers_element'
+import BuildSystemsElement from "src/ui/content_blocks/build_systems/build_systems_element"
+import ConferencesElement from "src/ui/content_blocks/conferences/conferences_element"
+import ArticlesElement from "src/ui/content_blocks/articles/articles_element"
+import LibrariesElement from "src/ui/content_blocks/libraries/libraries_element"
 import TabsElement from "src/ui/tabs/tabs_element"
